+++ src/r_things.h
@@ -209,8 +209,8 @@ extern INT32 numskins;
 extern INT32 numlocalskins;
 extern INT32 numallskins;
 extern skin_t skins[MAXSKINS];
-extern UINT8 skinstats[9][9][MAXSKINS];
-extern UINT8 skinstatscount[9][9];
+extern UINT16 skinstats[9][9][MAXSKINS];
+extern UINT16 skinstatscount[9][9];
 extern UINT16 skinsorted[MAXSKINS];
 
 void sortSkinGrid(void);
